Cache only when develop dud...

+++ library/tamal/tamal.php
@@ -65,7 +65,7 @@ class tamal extends prototype
     $php_file = APP_PATH.DS.'views'.DS.'cache'.DS.str_replace(TMP.DS, '', $file);
     $php_file = mkpath(dirname($php_file)).DS.basename($php_file);
 
-    if (is_file($php_file)) {
+    if (is_file($php_file) && (APP_ENV <> 'production')) {
       if (filemtime($file) > filemtime($php_file)) {
         unlink($php_file);
       }
